Move functions to traits
